web_creat/requirements.txt
streamlit
opencv-python
pillow
numpy
easyocr
requests